requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
langchain-core>=0.1.10

# Google Cloud Platform
google-cloud-firestore==2.13.1
google-cloud-tasks==2.14.2
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
google-auth==2.23.4

# AI & ML Services
google-generativeai==0.3.2
elevenlabs==0.2.26
openai==1.3.7

# Database & Caching
redis==5.0.1
firebase-admin==6.2.0

# Video Processing
ffmpeg-python==0.2.0

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Async & Background Tasks
celery==5.3.4
tenacity==8.2.3

# Observability
structlog>=23.0.0
# OpenTelemetry (선택적 - 필요시 주석 해제)
# opentelemetry-api>=1.20.0
# opentelemetry-sdk>=1.20.0
# opentelemetry-instrumentation-fastapi>=0.40b0
# opentelemetry-instrumentation-httpx>=0.40b0
# opentelemetry-exporter-gcp-trace>=1.9.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8 